+++ server/bingo/tests/test_friendship_model.py
@@ -1,0 +1,72 @@
+from django.db.utils import IntegrityError
+from django.test import TestCase
+from ..models import User, Friendship
+from django.core.exceptions import ValidationError
+
+
+class FriendshipTest(TestCase):
+    def setUp(self):
+        # Create test users
+        self.user1 = User.objects.create_user(
+            username="user1", email="user1@example.com", password="password")
+        self.user2 = User.objects.create_user(
+            username="user2", email="user2@example.com", password="password")
+        self.user3 = User.objects.create_user(
+            username="user3", email="user3@example.com", password="password")
+
+    def test_create_friendship(self):
+        # Tests whether a frienship can be created
+        friendship = Friendship.objects.create(
+            requester=self.user1, receiver=self.user2)
+        self.assertEqual(friendship.requester, self.user1)
+        self.assertEqual(friendship.receiver, self.user2)
+        # Checks default status is pending
+        self.assertEqual(friendship.status, Friendship.PENDING)
+
+    def test_unique_together_constraint(self):
+        # Create a friendship
+        Friendship.objects.create(requester=self.user1, receiver=self.user2)
+
+        # Attempt to create a duplicate friendship
+        with self.assertRaises(IntegrityError):
+            Friendship.objects.create(
+                requester=self.user1, receiver=self.user2)
+
+    def test_status_choices(self):
+        # Test creating a friendship with "accepted" status
+        friendship = Friendship.objects.create(
+            requester=self.user1, receiver=self.user2, status=Friendship.ACCEPTED)
+        self.assertEqual(friendship.status, Friendship.ACCEPTED)
+
+        # Test invalid status
+        with self.assertRaises(ValidationError):
+            Friendship.objects.create(
+                requester=self.user1, receiver=self.user3, status="invalid").full_clean()
+
+    def test_str_representation(self):
+        # Test the string representation of a friendship
+        friendship = Friendship.objects.create(
+            requester=self.user1, receiver=self.user2)
+        expected_str = f"Friend request from {
+            self.user1} to {self.user2} (Pending)"
+        self.assertEqual(str(friendship), expected_str)
+
+    def test_delete_user_cascades(self):
+        # Test that deleting a user cascades to related friendships
+        Friendship.objects.create(requester=self.user1, receiver=self.user2)
+        self.user1.delete()
+        self.assertEqual(Friendship.objects.count(), 0)
+
+    def test_reverse_friendship(self):
+        # Test whether two reverse friendship can be created
+        Friendship.objects.create(requester=self.user1, receiver=self.user2)
+
+        with self.assertRaises(ValidationError):
+            Friendship.objects.create(
+                requester=self.user2, receiver=self.user1).full_clean()
+
+    def test_friends_with_self(self):
+        # Test that a user cannot have a friendship with themselves
+        with self.assertRaises(ValidationError):
+            Friendship.objects.create(
+                requester=self.user1, receiver=self.user1).full_clean()
